+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Custom error message for when client update failed due to missing write permissions
 - Hide local source path in error messages (C:\Users\....)
 
+### Fixed
+- "Delete temp files" would delete the full download folder, deleting potential none-TNUC related files. Now it only deletes relevant files, and only deletes the folder if it's empty
+
 ## [1.26.0] - 2026-08-11
 This update focuses on community feedback from Discord. Make sure to join and let me know what you think about the new features, and if you have any (good) suggestions.
 - [Join community Discord](https://discord.gg/CSckkRu9Y5)